+++ models/ed/R/model2netcdf.ED2.R
@@ -134,20 +134,29 @@ model2netcdf.ED2 <- function(outdir,
     }
 
     # generate start/end dates for processing
-    if (y == start_year) {
-      start_date_real <- lubridate::ymd(start_date)
+    
+    # read_E_files already takes care of adjusting start and end dates to match
+    # ED2 output.  These adjustments dont' make sense for E files, so skip them.
+    if(rflag == "-E-"){
+      start_date_real <- start_date
+      end_date_real <- end_date
     } else {
-      #When would this be run?
-      start_date_real <- lubridate::make_date(y, 1, 1)
+      
+      if (y == start_year) {
+        start_date_real <- lubridate::ymd(start_date)
+      } else {
+        #When would this be run?
+        start_date_real <- lubridate::make_date(y, 1, 1)
+      }
+      
+      if (y == end_year) {
+        end_date_real <- lubridate::ymd(end_date)
+      } else {
+        #When would this be run?
+        end_date_real <- lubridate::make_date(y, 12, 31)
+      }
     }
-
-    if (y == end_year) {
-      end_date_real <- lubridate::ymd(end_date)
-    } else {
-      #When would this be run?
-      end_date_real <- lubridate::make_date(y, 12, 31)
-    }
-
+    
     # create lat/long nc variables
     lat <- ncdf4::ncdim_def("lat", "degrees_north",
                             vals = as.numeric(sitelat),
@@ -155,7 +164,7 @@ model2netcdf.ED2 <- function(outdir,
     lon <- ncdf4::ncdim_def("lon", "degrees_east",
                             vals = as.numeric(sitelon),
                             longname = "station_longitude")
-
+    
     # ----- put values to nc_var list
     nc_var <- list()
     for (i in seq_along(out_list)) {
@@ -937,9 +946,8 @@ put_T_values <-
 ##' area using the plant density and patch area before converting units to PEcAn
 ##' standards.
 ##'
-##' @param yr length 1 numeric vector; the year being processed
-##' @param yfiles numeric vector of the years on the filenames, will be used to
-##'   matched `h5_files` for that year
+##' @param yr unused. For consistency with `read_T_files()`.
+##' @param yfiles unused. For consistency with `read_T_files()`.
 ##' @param h5_files character vector of names of E h5 files (e.g.
 ##'   "analysis-E-1999-06-00-000000-g01.h5")
 ##' @param outdir directory where ED2 output files are found
@@ -971,100 +979,125 @@ read_E_files <- function(yr, yfiles, h5_files, outdir, start_date, end_date,
   stopifnot(!is.null(outdir), !is.null(start_date), !is.null(end_date), 
             !is.null(pfts))
   
-  # there are multiple -E- files per year
-  ysel <- which(yr == yfiles)
-  
-  # grab year-month info from file names, e.g. "199906"
-  times <- gsub(
-    "(.*)\\-(.*)\\-(.*)\\-(.*)\\-(.*)", "\\1\\2",
-    sapply(
-      strsplit(h5_files, "-E-"), 
-      function(x) x[2] # Select only the part of each name after res.flag
+  # #extract data from a single -E- .h5 file as a tibble
+  extract_E_file <- function(file) {
+    
+    # Cohort-level variables to extract
+    cohort_vars <- c(
+      "DBH", #diameter at breast height (cm) of each cohort
+      "AGB_CO", #cohort level above ground biomass (kgC/plant)
+      "MMEAN_NPPDAILY_CO", #net primary productivity (kgC/plant/yr) 
+      "MMEAN_TRANSP_CO", #Monthly mean leaf transpiration (kg/plant/s)
+      "BSEEDS_CO", #seed biomass in units of (kgC/plant)
+      "NPLANT", #plant density (plants/m2), required for /plant -> /m2 conversion
+      "PFT" #pft numbers
     )
-  )
-  
-  #Check that all the expected files exist using start_date and end_date
-  expected_ym <- seq(
-    lubridate::ymd(start_date),
-    # an E file is only written if a month is completed.
-    # E.g. start_date=2004-07-01, end_date=2004-08-31 will result in one E file for 2004-07
-    lubridate::floor_date(lubridate::ymd(end_date), "month") - lubridate::days(1), 
-    by = "month"
-  ) %>% format("%Y%m")
-  
-  if(!all(expected_ym %in% times)) {
-    #TODO: possibly not an error, but then need to use actual months of output files for time dimension in put_E_values(),  not start_date:end_date.
-    PEcAn.logger::logger.error("Not all expected E files found!")
+    
+    # Patch-level variables needed for calculations
+    patch_vars <- c(
+      "AREA", # fractional patch area relative to site area (unitless)
+      "AGE", #patch age since last disturbance
+      "PACO_N", #number of cohorts in each patch
+      "PACO_ID" #index of the first cohort of each patch.  Needed for figuring out which patch each cohort belongs to
+    )
+    
+    nc <- ncdf4::nc_open(file)
+    on.exit(ncdf4::nc_close(nc), add = FALSE)
+    avail_cohort <- cohort_vars[cohort_vars %in% names(nc$var)]
+    if(length(avail_cohort) == 0) {
+      PEcAn.logger::logger.warn("No cohort-level variables found!")
+      return(NULL)
+    }
+    avail_patch <- patch_vars[patch_vars %in% names(nc$var)]
+    if(length(avail_patch) == 0) {
+      PEcAn.logger::logger.warn("No patch-level variables found!")
+      return(NULL)
+    }
+    # Cohort-level and patch-level variables are extracted separately, then
+    # joined with a rolling-join
+    cohort_df <-
+      purrr::map(avail_cohort, function(.x) as.vector(ncdf4::ncvar_get(nc, .x))) %>%
+      purrr::set_names(avail_cohort) %>%
+      tibble::as_tibble() %>%
+      dplyr::mutate(COHORT_ID = 1:dplyr::n())
+    
+    patch_df <-
+      purrr::map(avail_patch, function(.x) as.vector(ncdf4::ncvar_get(nc, .x))) %>%
+      purrr::set_names(avail_patch) %>%
+      tibble::as_tibble() %>%
+      dplyr::mutate(PATCH_ID = 1:dplyr::n())
+    
+    # join patch data with cohort data.  PACO_N is the number of cohorts (i.e.
+    # rows) in each patch, so I use that to figure out which rows are in which
+    # patch.
+    cohort_patch_id <- c()
+    for (i in seq_len(nrow(patch_df))) {
+      cohort_patch_id <- c(cohort_patch_id, rep(i, patch_df$PACO_N[i]))
+    }
+    cohort_df$PATCH_ID <- cohort_patch_id
+    
+    # Then the cohort and patch dataframes can be joined by PATCH_ID
+    dplyr::left_join(cohort_df, patch_df, by = "PATCH_ID") %>% 
+      # Dates in filename are not valid because day is 00.  Extract just year
+      # and month and use lubridate:ym() to build date
+      dplyr::mutate(date = stringr::str_match(basename(file), "(\\d{4}-\\d{2})-\\d{2}")[, 2] %>% lubridate::ym()) %>%
+      dplyr::select(date, dplyr::everything()) %>%
+      dplyr::select(-"PACO_ID")
   }
   
-  # lets make it work for a subset of vars fist
-  # TODO :  read all (or more) variables, functionality exists, see below
-  varnames <-
-    c(
-      "DBH", #diameter at breast height (cm)
-      "DDBH_DT", #change in DBH (cm/plant/yr) 
-      "AGB_CO", #cohort level above ground biomass (kgC/plant)
-      "MMEAN_NPPDAILY_CO", #net primary productivity (kgC/m2/yr)
-      "MMEAN_TRANSP_CO", #Monthly mean leaf transpiration (kg/m2/s)
-      "BSEEDS_CO", #seed biomass in units of (kgC/plant)
-      "NPLANT" #plant density (plants/m2), required for /plant -> /m2 conversion
+  # Extract data from all the .h5 files of the yr
+  raw <- 
+    purrr::map(file.path(outdir, h5_files[yr == yfiles]), extract_E_file) %>% 
+    dplyr::bind_rows() 
+  
+  # Unit conversions
+  out <- raw %>% 
+    dplyr::mutate(
+      # Convert units of per plant to per m^2 by multiplying by plant density
+      # (plant/m^2)
+      dplyr::across(
+        dplyr::any_of(c("BSEEDS_CO", "AGB_CO", "MMEAN_NPPDAILY_CO", "MMEAN_TRANSP_CO")),
+        function(.x) .x * .data$NPLANT
+      ),
+      # Convert to pecan standard units
+      MMEAN_NPPDAILY_CO = PEcAn.utils::ud_convert(.data$MMEAN_NPPDAILY_CO, "kg/m^2/yr", "kg/m^2/s")
+    ) %>%
+    # Weighted summary
+    # For each month (-E- file) and PFT...
+    dplyr::group_by(.data$date, .data$PFT) %>% 
+    # ... get a weighted sum across cohorts and patches
+    dplyr::summarize(
+      dplyr::across(
+        dplyr::any_of(c("BSEEDS_CO", "AGB_CO", "MMEAN_NPPDAILY_CO", "MMEAN_TRANSP_CO", "NPLANT")),
+        function(.x) sum(.x * .data$AREA)
+      ),
+      #(or a mean when it makes sense)
+      DBH_mean = mean(.data$DBH),
+      .groups = "drop"
     )
   
-  # List of vars to extract includes the requested one, plus others needed below 
-  vars <- c(
-    varnames,
-    "PFT"#, #pft numbers
-
-    # "AREA", #patch area relative to site area (unitless)
-    # "AREA_SI", #site area relative to polygon area (unitless)
-    # "PACO_N" #number of cohorts in each patch
-    )
+  # Rename variables to PEcAn standard and convert to list
+  n_pft <- length(unique(out$PFT))
+  out_list <-
+    out %>% 
+    dplyr::arrange(.data$PFT, .data$date) %>% 
+    dplyr::select(-"PFT", -"date") %>% 
+    #output is expected to be list of matrixes with ncol == number of PFTs.
+    #Here, I make a tibble with matrix-columns (each data frame column is a
+    #n_pft-wide matrix), then convert it to a list.
+    dplyr::summarize(dplyr::across(dplyr::everything(),
+      function(.x) matrix(.x, ncol = n_pft)
+    )) %>% 
+    dplyr::select(
+      #PEcAn name    #ED2 name
+      "AGB_PFT"    = "AGB_CO",
+      "BSEEDS"     = "BSEEDS_CO",
+      "DBH"        = "DBH_mean",
+      "NPP_PFT"    = "MMEAN_NPPDAILY_CO",
+      "TRANSP_PFT" = "MMEAN_TRANSP_CO",
+      "DENS"       = "NPLANT"
+    ) %>% as.list()
   
-  # list to collect outputs
-  ed.dat <- list()
-  
-  # loop over the files for that year
-  for(i in ysel) {
-    
-    nc <- ncdf4::nc_open(file.path(outdir, h5_files[i]))
-    on.exit(ncdf4::nc_close(nc), add = FALSE)
-    allvars <- names(nc$var)
-    if (!is.null(vars)) allvars <- allvars[ allvars %in% vars ]
-    
-    # extract all the data into list
-    #TODO warn if a variable isn't available and return -999 (?)
-    if (length(ed.dat) == 0){
-      for (j in 1:length(allvars)) {
-        ed.dat[[j]] <- list()
-        ed.dat[[j]][[1]] <- ncdf4::ncvar_get(nc, allvars[j])
-      }
-      names(ed.dat) <- allvars
-    } else {
-      
-      # 2nd and more months
-      t <- length(ed.dat[[1]]) + 1
-      
-      for (j in 1:length(allvars)) {
-        
-        k <- which(names(ed.dat) == allvars[j])
-        
-        if (length(k)>0) {
-          
-          ed.dat[[k]][[t]] <- ncdf4::ncvar_get(nc, allvars[j])
-          
-        } else { ## add a new ed.datiable. ***Not checked (shouldn't come up?)
-          
-          ed.dat[[length(ed.dat)+1]] <- list()    # Add space for new ed.datiable
-          ed.dat[[length(ed.dat)]][1:(t-1)] <- NA # Give NA for all previous time points
-          ed.dat[[length(ed.dat)]][t] <- ncdf4::ncvar_get(nc, allvars[j]) # Assign the value of the new ed.datiable at this time point
-          names(ed.dat)[[length(ed.dat)]] <- allvars[j] 
-          
-        }
-      }      
-    }
-    
-  } # end ysel-loop
-
   
   # even if this is a SA run for soil, currently we are not reading any variable
   # that has a soil dimension. "soil" will be passed to read.output as pft.name
@@ -1075,72 +1108,11 @@ read_E_files <- function(yr, yfiles, h5_files, outdir, start_date, end_date,
     #TODO: print a message??
     pfts <- pfts[!(soil.check)]
   }
+  out_list$PFT <- sort(pfts) #named vector for matching PFT numbers to names
   
-  # Aggregate over PFT and DBH bins  
-  for (i in seq_along(ysel)) {
- 
-    #At this point, every element in ed.dat is a list of variables each having
-    #one element per month that is usually a vector but sometimes a matrix. For
-    #each variable, the following needs to be mapped to to each month:
-    #' 1) Is the variable a matrix?  If so, get colsums to turn it into a vector with one element per cohort
-    #' 2) Is the variable in per-plant units? If so, it needs converting to per area units
-    #' 3) Are the units PEcAn standard? If not, they need converting (e.g. with PEcAn.utils::ud_convert()) (make use of PEcAn.utils::standard.vars?)
-    #' 4) group by PFT and sum cohorts
-    
-    #TODO: This is written in a way that only apply to cohort-level variables.  Should this be generalized?
-    out <- 
-      #TODO, this outer imap could be made into a for-loop if it makes it easier for people to read and edit in the future.  not necessarily faster with imap
-      purrr::imap(ed.dat[names(ed.dat) %in% varnames], ~{ 
-        #.x is elements of ed.dat and .y is names of ed.dat
-        #1) collapse matrix variables into vector
-        if (all(purrr::map_lgl(.x, is.matrix))) {
-          var <- purrr::map(.x, colSums)
-        } else {
-          var <- .x
-        }
-        
-        #2) do per plant -> per area correction
-        if (.y %in% c("BSEEDS_CO", "AGB_CO")) {
-          var <- purrr::map2(var, ed.dat$NPLANT, `*`)
-        }
-        
-        #3) convert units to PEcAn standard if necessary
-        #input units are according to the ED2 source code: https://raw.githubusercontent.com/EDmodel/ED2/master/ED/src/memory/ed_state_vars.F90, output units are according to PEcAn.utils::standard_vars
-        if(.y == "MMEAN_NPPDAILY_CO") {
-          var <- purrr::map(var, ~ PEcAn.utils::ud_convert(.x, u1 = "kg/m2/yr", u2 = "kg/m2/s"))
-        }
-
-        var
-      }) 
-    
-    #sum cohorts by PFT
-    out <- 
-      purrr::map(.x = out, #for each variable in `out`
-          ~purrr::map2(.x = .x, .y = ed.dat$PFT, #for each month in each variable
-                ~ tapply(.x, .y, sum) #sum variable by PFT number
-          ))
-  }
-    
-  #Bind rows for months together to produce a matrix with ncol = length(pfts) and nrow = number of months
-  out <- purrr::map(out, ~do.call(rbind, .x))
+  return(out_list)
   
-  out$PFT <- pfts #named vector for matching PFT numbers to names
-  
-  #New varnames to match PEcAn standard
-  names(out) <- dplyr::case_when(
-                  #ED2 name             #PEcAN name
-    names(out) == "AGB_CO"            ~ "AGB_PFT",
-    names(out) == "BSEEDS_CO"         ~ "BSEEDS",
-    names(out) == "DDBH_DT"           ~ "DDBH",
-    names(out) == "MMEAN_NPPDAILY_CO" ~ "NPP_PFT",
-    names(out) == "MMEAN_TRANSP_CO"   ~ "TRANSP_PFT",
-    names(out) == "NPLANT"            ~ "DENS",
-    TRUE ~ names(out)
-  )
-  
-  return(out)
-  
-} # end read_E_files
+}
 
 
 
@@ -1196,13 +1168,28 @@ put_E_values <-
   ##### setup output time and time bounds
   ## Create a date vector that contains each month of the model run (e.g.
   ## "2001-07-01" "2001-08-01" "2001-09-01"....) and which is the correct length
-  ## for each full or partial year
+  ## for each full or partial year.  In the last year of the simulation, the .h5
+  ## for the  month in end_date will never be written, because it won't be
+  ## complete.
+  if(yr == lubridate::year(start_date)){
+    date_vec_start <- lubridate::floor_date(lubridate::ymd(start_date), "month")
+    #TODO I should check what happens if a run starts not on the first day of the month
+  } else {
+    date_vec_start <- lubridate::make_date(year = yr, month = 1, day = 1)
+  }
+  
+  date_vec_end <- lubridate::floor_date(lubridate::ymd(end_date), "month")
+  if (yr == lubridate::year(end_date)) {
+    #remove the last month
+    date_vec_end <- date_vec_end - months(1)
+  } else {
+    date_vec_end <- lubridate::make_date(year = yr, month = 12, day = 1)
+  }
+  
   output_date_vector <-
     seq(
-      lubridate::ymd(start_date),
-      # an E file is only written if a month is completed.
-      # E.g. start_date=2004-07-01, end_date=2004-08-31 will result in one E file for 2004-07
-      lubridate::floor_date(lubridate::ymd(end_date), "month") - lubridate::days(1), 
+      date_vec_start,
+      date_vec_end, 
       by = "month"
     )
   ## Create a vector of the number of days in each month by year (e.g. 31 31 30
@@ -1287,13 +1274,6 @@ put_E_values <-
       dim = list(lon, lat, t, p),
       missval = -999,
       longname = "Diameter at breast height by PFT"
-    ),
-    ncdf4::ncvar_def(
-      "DDBH", #original ED2 name: DDBH_DT
-      units = "cm yr-1",
-      dim = list(lon, lat, t, p),
-      missval = -999,
-      longname = "Rate of change in dbh by PFT"
     ),
    ncdf4::ncvar_def(
      "NPP_PFT", #original ED2 name: MMEAN_NPPDAILY_CO